--- src/app/tv-series/tv-series.component.html
@@ -1,3 +1,0 @@
-<p>
-  tv-series works!
-</p>
