--- src/app/tv-series/tv-series.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-tv-series',
-  templateUrl: './tv-series.component.html',
-  styleUrls: ['./tv-series.component.css']
-})
-export class TvSeriesComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
